+++ certlogic/certlogic-js/CHANGELOG.md
@@ -7,6 +7,7 @@
   This replaces the semantics where a desugared `"or"` always returned a boolean value, due to the use of [De Morgan's laws](https://en.wikipedia.org/wiki/De_Morgan%27s_laws).
   (Also: an `"or"` with a single operand reduces to that operand.)
 * Build improvement: perform deduplication of (nested) dependencies.
+* Introduce the concept of “boolsiness” (internally) by means of a type `Boolsiness` and a function `boolsiness`.
 
 
 ## 1.1.2